+++ src/main/java/com/game/enchanter/entities/Wall.java
@@ -10,7 +10,7 @@ public class Wall extends Obstacle implements Renderable {
 	
 	private static final float[] COLOR = { 0.0f, 0.0f, 1.0f }; // blue
 
-	public Wall(float _x, float _y, float _width, float _height, boolean _passable) {
+	public Wall(int _x, int _y, int _width, int _height, boolean _passable) {
 		super(_x, _y, _width, _height, _passable);
 	}
